+++ ProyectoRE/src/main/java/com/ProyectoRE/dao/FavoritoDao.java
@@ -5,6 +5,6 @@ import com.ProyectoRE.domain.Favorito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface FavoritoDao extends JpaRepository<Favorito, Long> {
+public interface FavoritoDao extends JpaRepository<Favorito, Integer> {
     
 }
