+++ ProyectoRE/src/main/java/com/ProyectoRE/dao/ResenasDao.java
@@ -5,6 +5,6 @@ import com.ProyectoRE.domain.Resenas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ResenasDao extends JpaRepository<Resenas, Long> {
+public interface ResenasDao extends JpaRepository<Resenas, Integer> {
     
 }
