+++ ProyectoRE/src/main/java/com/ProyectoRE/dao/ServiciosPremiumDao.java
@@ -5,6 +5,6 @@ import com.ProyectoRE.domain.ServiciosPremium;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ServiciosPremiumDao extends JpaRepository<ServiciosPremium, Long> {
+public interface ServiciosPremiumDao extends JpaRepository<ServiciosPremium, Integer> {
     
 }
