+++ ProyectoRE/src/main/java/com/ProyectoRE/dao/TipoPropiedadDao.java
@@ -5,6 +5,6 @@ import com.ProyectoRE.domain.TipoPropiedad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface TipoPropiedadDao extends JpaRepository<TipoPropiedad, Long> {
+public interface TipoPropiedadDao extends JpaRepository<TipoPropiedad, Integer> {
     
 }
