+++ ProyectoRE/src/main/java/com/ProyectoRE/dao/UsuarioDao.java
@@ -5,6 +5,6 @@ import com.ProyectoRE.domain.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface UsuarioDao extends JpaRepository<Usuario, Long> { 
+public interface UsuarioDao extends JpaRepository<Usuario, Integer> { 
     
 }
